Init tennis utils for converting score to tennis score
