+++ TennisScoreTable/src/test/java/com/yeahmama/tennis/score/TennisGameScoreTest.java
@@ -88,54 +88,6 @@ public class TennisGameScoreTest {
     }
 
     @Test
-    public void should_give_0_as_tennis_score_when_score_is_0() {
-        // Arrange
-        int score = 0;
-
-        // Act
-        String tennisScore = tennisGameScore.toTennisScore(score);
-
-        // Assert
-        assertThat(tennisScore).isEqualTo("0");
-    }
-
-    @Test
-    public void should_give_15_as_tennis_score_when_score_is_1() {
-        // Arrange
-        int score = 1;
-
-        // Act
-        String tennisScore = tennisGameScore.toTennisScore(score);
-
-        // Assert
-        assertThat(tennisScore).isEqualTo("15");
-    }
-
-    @Test
-    public void should_give_30_as_tennis_score_when_score_is_2() {
-        // Arrange
-        int score = 2;
-
-        // Act
-        String tennisScore = tennisGameScore.toTennisScore(score);
-
-        // Assert
-        assertThat(tennisScore).isEqualTo("30");
-    }
-
-    @Test
-    public void should_give_40_as_tennis_score_when_score_is_3() {
-        // Arrange
-        int score = 3;
-
-        // Act
-        String tennisScore = tennisGameScore.toTennisScore(score);
-
-        // Assert
-        assertThat(tennisScore).isEqualTo("40");
-    }
-
-    @Test
     public void should_add_game_over_column_to_game_score_table_when_it_has_a_game_winner() {
         // Arrange
         tennisGameScore.addPointPlayerOne();
